# server/src/Tgm.Roborally.Server/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "contact" : {
      "email" : "nbrugger@student.tgm.ac.at",
      "name" : "Nils Brugger"
    },
    "description" : "This api controlls the flow of a game and provides it's data.\nIt is desiged to be RESTfull so the structure works simmilar as file system.\nThe service will run and only work in a local network, `game.host` is the IP of the Computer hosting the game and will be found via a IP scan",
    "license" : {
      "name" : "Apache 2.0"
    },
    "title" : "Robot Rally Game logic engine",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "description" : "Production",
    "url" : "http://game.host/v1"
  }, {
    "description" : "Debug",
    "url" : "http://localhost:5050/v1"
  } ],
...
  "paths" : {
...
    "/games/{game_id}/upgrades/shop" : {
      "get" : {
        "description" : "Retuns a list of all cards in the upgrade shop",
        "operationId" : "getUpgradeShop",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UpgradeShop"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Get upgrade shop",
        "tags" : [ "Upgrades" ]
      },
      "patch" : {
        "description" : "Buy this Upgrade from the shop.\n### Note\nIf your robot allready owns 3 Upgrades you have to use the `exchange` parameter. This is the ID of the Upgrade to replace with the bought one",
        "operationId" : "buyUpgrade",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "upgrade",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UpgradeID"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "exchange",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/UpgradeID"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Buy Upgrade",
        "tags" : [ "Upgrades" ]
      }
    },
    "/games/{game_id}/upgrades/{upgrade_id}" : {
      "get" : {
        "description" : "Get detailed information about the Upgrade",
        "operationId" : "getUpgradeInformation",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "upgrade_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UpgradeID"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Upgrade"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Get upgrade information",
        "tags" : [ "Upgrades" ]
      }
    },
    "/games/{game_id}/upgrades/" : {
      "get" : {
        "description" : "Returns a list of all Upgrade IDs",
        "operationId" : "getAllUpgradeIDs",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/UpgradeID"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "get all Upgrades",
        "tags" : [ "Upgrades" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "GameID" : {
        "description" : "**Unique**<br>\nThis is the parameter a game is identified by",
        "example" : 537,
        "format" : "int32",
        "maximum" : 2048,
        "minimum" : 0,
        "title" : "GameID",
        "type" : "integer"
      },
...
      "Error-Message" : {
        "description" : "A message about that went wrong. Usefull to display users a short and usefull prompt",
        "properties" : {
          "message" : {
            "description" : "A short message describing what happened in human words",
            "type" : "string"
          },
          "error" : {
            "description" : "The error/exception",
            "type" : "string"
          }
        },
        "required" : [ "message" ],
        "title" : "Error-Message",
        "type" : "object"
      },
...
      "UpgradeShop" : {
        "description" : "A shop to buy upgrades from. Each element is buyable once and then vanishes from the list",
        "example" : {
          "information" : {
            "open" : true
          },
          "upgrades" : [ null, null, null, null, null ]
        },
        "properties" : {
          "upgrades" : {
            "items" : {
              "$ref" : "#/components/schemas/UpgradeID"
            },
            "maxItems" : 10,
            "minItems" : 3,
            "title" : "UpgradeShop",
            "type" : "array"
          },
          "information" : {
            "$ref" : "#/components/schemas/UpgradeShopInformation"
          }
        },
        "type" : "object"
      },
      "Upgrade" : {
        "additionalProperties" : false,
        "description" : "A upgrade is a module making a robot stronger",
        "example" : {
          "cost" : 0,
          "permanent" : true,
          "values" : [ {
            "name" : "name",
            "value" : 1
          }, {
            "name" : "name",
            "value" : 1
          }, {
            "name" : "name",
            "value" : 1
          }, {
            "name" : "name",
            "value" : 1
          }, {
            "name" : "name",
            "value" : 1
          } ],
          "name" : "Penetration Lazer Mk.2",
          "description" : "null",
          "id" : 5962,
          "rounds" : 6
        },
        "minProperties" : 4,
        "properties" : {
          "cost" : {
            "default" : 2,
            "description" : "The energy cost to buy this upgrade",
            "maximum" : 5,
            "minimum" : 0,
            "type" : "integer"
          },
          "name" : {
            "description" : "The ame to display for this Upgrade. ***Not*** unique (identifying)",
            "example" : "Penetration Lazer Mk.2",
            "maxLength" : 27,
            "minLength" : 2,
            "type" : "string"
          },
          "permanent" : {
            "default" : true,
            "description" : "if true the card belongs to the player for the rest of the game",
            "type" : "boolean"
          },
          "description" : {
            "default" : "null",
            "description" : "A description about the effect of the card.\nVariables are using the format `{name}` where *name* refers to the names in `values`. ",
            "maxLength" : 300,
            "type" : "string"
          },
          "rounds" : {
            "description" : "If the Upgrade is not permanent this variable defines the number of rounds this Upgrade works",
            "maximum" : 10,
            "minimum" : 1,
            "type" : "integer"
          },
          "values" : {
            "description" : "Defines number values for the upgrade.<br>Example:\nEffect: \"You have {registers} additonal Registers\"<br> `{registers}` is the number of the regsiters (that will be added) and the exact value will be defined in here (`values`)",
            "items" : {
              "$ref" : "#/components/schemas/Pair"
            },
            "maxItems" : 10,
            "type" : "array",
            "uniqueItems" : true
          },
          "type" : {
            "$ref" : "#/components/schemas/UpgradeType"
          },
          "id" : {
            "description" : "The id of an upgrade. **Unique**",
            "maximum" : 10000,
            "minimum" : 0,
            "title" : "UpgradeID",
            "type" : "integer"
          }
        },
        "required" : [ "cost", "name", "type" ],
        "title" : "Upgrade",
        "type" : "object"
      },
...
      "Pair" : {
        "additionalProperties" : false,
        "description" : "A named value. A pair of an index and value. (Part of a map)",
        "example" : {
          "name" : "name",
          "value" : 1
        },
        "maxProperties" : 2,
        "minProperties" : 1,
        "properties" : {
          "name" : {
            "description" : "The name the value is bound to",
            "minLength" : 1,
            "type" : "string"
          },
          "value" : {
            "description" : "The value behind the name. Can be null",
            "type" : "integer"
          }
        },
        "required" : [ "name" ],
        "title" : "Pair",
        "type" : "object"
      },
      "UpgradeType" : {
        "description" : "Defines the type (the code/actions) this card will have\n* `generator` : Generates *x* energy every round",
        "enum" : [ "generator" ],
        "maxLength" : 300,
        "title" : "UpgradeType",
        "type" : "string"
      },
...
      "UpgradeID" : {
        "description" : "The id of an upgrade. **Unique**",
        "maximum" : 10000,
        "minimum" : 0,
        "title" : "UpgradeID",
        "type" : "integer"
      },
      "UpgradeShopInformation" : {
        "example" : {
          "open" : true
        },
        "properties" : {
          "open" : {
            "default" : true,
            "description" : "if true you can buy uprades",
            "type" : "boolean"
          }
        },
        "title" : "UpgradeShopInformation",
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "Host-token-access" : {
        "description" : "The Host of the game (and of the API) has an unique identificaton granting him game manipulating abilities. By this means it is also possible to have a host and a client at the same machine. \n> This could be used for AI`s (bot players)",
        "in" : "query",
        "name" : "hid",
        "type" : "apiKey"
      },
      "Player-Access-Token" : {
        "description" : "Players need a token to join a lobby/game. If the host sets a password the player needs the password to get the token",
        "in" : "query",
        "name" : "pat",
        "type" : "apiKey"
      }
    }
  }
}